fix: Fix chatting bugs
