+++ build.gradle.kts
@@ -6,7 +6,7 @@ plugins {
 }
 
 group = "xyz.netherald"
-version = "2.0.1"
+version = "2.1.0"
 
 repositories {
     mavenCentral()
